requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
async-generator==1.10
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.40.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
contourpy==1.2.0
customtkinter==5.0.5
cycler==0.12.1
darkdetect==0.8.0
datasets==2.16.1
dill==0.3.7
distro==1.9.0
docutils==0.20.1
etelemetry==0.3.1
exceptiongroup==1.2.0
fastapi==0.108.0
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.13.1
fitz==0.0.1.dev2
fonttools==4.47.0
frontend==0.0.3
frozenlist==1.4.1
fsspec==2023.10.0
future==0.18.3
gradio==4.13.0
gradio_client==0.8.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
imageio==2.25.1
imageio-ffmpeg==0.4.8
importlib-resources==6.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
looseversion==1.3.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
netifaces==0.11.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.3
openai==1.12.0
opencv-python==4.7.0.68
orjson==3.9.10
outcome==1.2.0
packaging==23.2
pandas==2.1.4
pathlib==1.0.1
pefile==2023.2.7
peft==0.4.0
pillow==10.2.0
proglog==0.1.10
prov==2.0.0
psutil==5.9.8
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
pydub==0.25.1
Pygments==2.17.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyMySQL==1.1.0
pyparsing==3.1.1
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
pywin32-ctypes==0.2.2
pyxnat==1.6
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
safetensors==0.4.2
scapy==2.5.0
scipy==1.11.4
selenium==4.8.0
semantic-version==2.10.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.32.0.post1
sympy==1.12
tokenizers==0.13.3
tomlkit==0.12.0
toolz==0.12.0
torch==2.2.0
tqdm==4.66.1
traits==6.3.2
transformers==4.31.0
trio==0.22.0
trio-websocket==0.9.2
trl==0.4.7
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.25.0
websockets==11.0.3
wget==3.2
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
